fix:edited the create batchscreen
